--- src/org/usfirst/frc/team4571/robot/CalculationConstants.java
@@ -1,10 +1,0 @@
-package org.usfirst.frc.team4571.robot;
-
-/**
- * 
- * @author arjunrao87
- *
- */
-public final class CalculationConstants {
-
-}
