--- src/org/usfirst/frc/team4571/robot/HardwareConfiguration.java
@@ -1,5 +1,0 @@
-package org.usfirst.frc.team4571.robot;
-
-public class HardwareConfiguration {
-
-}
