+++ src/org/usfirst/frc/team4571/robot/computations/ComputationConstants.java
@@ -1,0 +1,10 @@
+package org.usfirst.frc.team4571.robot.computations;
+
+/**
+ * 
+ * @author arjunrao87
+ *
+ */
+public final class ComputationConstants {
+
+}
