+++ src/org/usfirst/frc/team4571/robot/subsystems/ElevatorSubsystem.java
@@ -1,6 +1,6 @@
 package org.usfirst.frc.team4571.robot.subsystems;
 
-import org.usfirst.frc.team4571.robot.ElevatorSpeedController;
+import org.usfirst.frc.team4571.robot.components.ElevatorSpeedController;
 
 import edu.wpi.first.wpilibj.command.Subsystem;
